Fix courier system; Add ember_spirit; (#47)
Forbid silence ability usage on item channeling enemies (meteor_hammer);
Change juggernaut ability level up;
Fix troll_warlord ultimate;
Forbid location ability usage on invulnerable units (ember_spirit, puck);

Fix arc_warden GetBot;
Fix bane, beastmaster, centaur, hoodwink abilities;
Decrease the frequency of bounty_hunter, treant querying enemies 1600 away from npcBot to reduce warning in the console;
Change chen, bounty_hunter ability;
Restrict occasions for casting winter_wyvern_cold_embrace;
Add puck draft;
Change CourierSystem.lua to let every bot get access to the courier;
Change some item purchase solution to save their money;

Commit local dictionary for dota vocabulary

Fix brewmaster cannot move while primal_split;

Use glyph when target is quickly losing health;
Forbid furion use sprout at himself while invisible and not attacked;
Forbid obsidian_destroyer, shadow_demon use imprisonment ability on severely disabled enemies;
Add beastmaster_hawk_bomb_dive;
Try not to use tango at trees inside enemy towers' attack range;

Add ember_spirit;
Add beastmaster_hawk_dive minion use;
Improve templar_assassin_psionic_trap;
Fix bots stuck because of repeatedly using ring_of_basilius or buckler;

Add pudge use rot when channeling;
Add shadow_demon_soul_catcher, provide shadow_demon better cast target of disruption and demonic_purge;

Fix newly-added glyph usage;

Fix drow_ranger use split_shot at target;
Fix bugs in omniknight shard ability;
Fix bugs in item purchase when reloading;
Remove consumable purchase for heros who want to buy bottle;
Enable skeleton_king buy Aghanim's shard when antimage is on enemy team;

Fix abilities in ember_spirit, pudge, hoodwink;

Recall courier when hero's inventory is full;

Fix tipos in change log

Add tide_hunter, magnataur cancel ultimate when there are no enemies around;

Forbid enchantress, drow_ranger, obsidian_destroyer, silencer cast attack ability on ethereal or magic immune targets or when totally blinded;
Fix bug when lich try to cast frost_armour on enemy;

Add armlet of mordiggian usage;
Don't heal target under ancient_apparition_ice_blast (including armlet and power treads);
Fix bugs in chen, leshrac, and omniknight;

Add bane cancel nightmare;

Fix errors when npcMostDangerousEnemy is unassigned;
Add tusk use snowball with friends and scroll before timeout;
Forbid winter_wyvern_cold_embrace cast on friends with time sensitive modifiers (e.g. bkb, fv_chronosphere), applies to snowball as well;

Add ability information of basic levels;
Antimage, ember_spirit now only dodge dodge-worth abilities (stun, strong disable, massive damage);
Add bane try to cancel brain_snap and ult when target has linken's sphere, lotus_orb, or antimage_counterspell active;
Heroes can use any blink (agi/str/int);
Add ursa use enrage when disabled;
Remove a pair of unused local;

Fix errors in magnataur, sniper;

Fix incorrect scepter usage of chen, clinkz, doom, enchantress;
Cancel axe_culling_blade when target gains health suddenly;

Add alchemist throw concoction earlier when enemies have stun or silence;
Improve antimage blink location when attacking enemies;
Improve winter_wyvern_cold_embrace target;
Temporarily remove all buckler and ring_of_basilius purchase;

Add announcement of discord;

Fix spectre try to use spectral_dagger at creeps;
Fix some errors in earth_spirit;

Avoid using ultimate abilities at non-hero units;
Add unerlord_dark_rift cancel and scepter;
Add alchemist use scepter;
[dev-only] Add coroutine-style consider call;
Add life_stealer scepter;
Add lina shard;
Better obsidian_destroyer_imprisonment behavior at allies;

Add fear, hynopsis, root detect, and forbid certain abilities at these negative modifiers;

Fix arc_warden_tempest_double try to use abilities while mana is not enough;
Fix treant_natures_grasp use at target unit (can only target location);

+++ ability_item_usage_disruptor.lua
@@ -469,6 +469,14 @@ Consider[4]=function()
 	--------------------------------------
 	-- Global high-priorty usage
 	--------------------------------------
+	local channelingEnemies = AbilityExtensions:Filter(enemys, function(t)
+		return CanCast[4](t) and AbilityExtensions:IsChannelingAbility(t) and AbilityExtensions:IsChannelingBreakWorthAbility(t)
+	end)
+	channelingEnemies = AbilityExtensions:Max(channelingEnemies, function(t) return t:GetCurrentActiveAbility():GetCooldown() end)
+	if channelingEnemies ~= nil and channelingEnemies[1]:GetCurrentActiveAbility():GetCooldown() >= ability:GetCooldown() * 0.7 then
+		return BOT_ACTION_DESIRE_MODERATE, AbilityExtensions:FindAOELocationAtSingleTarget(npcBot, channelingEnemies, Radius, CastRange, CastPoint)
+	end
+
 	--try to kill enemy hero
 	if(npcBot:GetActiveMode() ~= BOT_MODE_RETREAT ) 
 	then
@@ -536,11 +544,11 @@ Consider[4]=function()
 			return BOT_ACTION_DESIRE_LOW, locationAoE.targetloc;
 		end
 	
-		local npcEnemy = npcBot:GetTarget();
+		local npcEnemy = AbilityExtensions:GetTargetIfGood(npcBot)
 
 		if ( npcEnemy ~= nil ) 
 		then
-			if ( CanCast[abilityNumber]( npcEnemy ) )
+			if CanCast[abilityNumber]( npcEnemy ) and AbilityExtensions:IsNonIllusionHero(npcEnemy)
 			then
 				return BOT_ACTION_DESIRE_HIGH, npcEnemy:GetExtrapolatedLocation(CastPoint);
 			end
